chore: Renaming of importer and extracted new wrapper builder
