+++ tests/dike/surroundings/wrapper/test_surroundings_wrapper.py
@@ -4,9 +4,16 @@ from typing import Callable, Iterator
 import pytest
 from shapely.geometry import Point
 
-from koswat.configuration.io.ini.koswat_general_ini_fom import SurroundingsSectionFom
+from koswat.configuration.io.ini.koswat_general_ini_fom import (
+    InfrastructureSectionFom,
+    SurroundingsSectionFom,
+)
 from koswat.configuration.io.koswat_surroundings_importer import (
-    KoswatSurroundingsWrapperCollectionBuilder,
+    SurroundingsWrapperCollectionImporter,
+)
+from koswat.configuration.settings.koswat_general_settings import (
+    InfraCostsEnum,
+    SurtaxFactorEnum,
 )
 from koswat.dike.surroundings.point.point_surroundings import PointSurroundings
 from koswat.dike.surroundings.wrapper.infrastructure_surroundings_wrapper import (
@@ -31,148 +38,66 @@ class TestSurroundingsWrapper:
             InfrastructureSurroundingsWrapper,
         )
 
-    # @pytest.mark.parametrize(
-    #     "obstacle_name",
-    #     [pytest.param("buildings"), pytest.param("railways"), pytest.param("waters")],
-    # )
-    # def test_set_obstacles_polderside(
-    #     self,
-    #     obstacle_name: str,
-    #     distances_to_surrounding_point_builder: Callable[
-    #         [Point, list[float]], PointSurroundings
-    #     ],
-    # ):
-    #     # 1. Define test data.
-    #     _obstacles_polderside = SurroundingsObstacle()
-    #     _locations = [
-    #         Point(4.2, 2.4),
-    #         Point(4.2, 4.2),
-    #         Point(2.4, 2.4),
-    #     ]
-    #     _obstacles_polderside.points = list(
-    #         map(distances_to_surrounding_point_builder, _locations, [[0], [2], [24]])
-    #     )
+    @pytest.fixture(name="acceptance_surroundings_wrapper_fixture")
+    def _get_surroundings_wrapper_with_infrastructure_fixture(
+        self, request: pytest.FixtureRequest
+    ) -> Iterator[SurroundingsWrapper]:
+        # Shp locations file
+        _shp_file = test_data.joinpath("acceptance", "shp", "dike_locations.shp")
+        assert _shp_file.is_file()
 
-    #     # 2. Run test.
-    #     _surroundings = SurroundingsWrapper(
-    #         **{
-    #             f"apply_{obstacle_name}": True,
-    #             f"{obstacle_name}_polderside": _obstacles_polderside,
-    #         }
-    #     )
+        # Surroundings directory
+        _surroundings_analysis_path = test_data.joinpath(
+            "acceptance", "surroundings_analysis", "10_1"
+        )
+        assert _surroundings_analysis_path.is_dir()
 
-    #     # 3. Verify expectations.
-    #     assert isinstance(_surroundings.buildings_polderside, SurroundingsObstacle)
+        # Create a dummy dir to avoid importing unnecessary data.
+        _dir_name = get_test_results_dir(request)
+        _temp_dir = test_results.joinpath(_dir_name, "10_1")
+        if _temp_dir.exists():
+            shutil.rmtree(_temp_dir)
+        shutil.copytree(_surroundings_analysis_path, _temp_dir)
 
-    #     # Doing it like this because sorting requires setting up a sorting key.
-    #     _explored = []
-    #     assert len(set(_surroundings.obstacle_locations)) == len(_locations)
-    #     for _sp in _surroundings.obstacle_locations:
-    #         assert _sp.location in _locations
-    #         # Check there are no repeated points, thus invalidating the test.
-    #         assert _sp not in _explored
-    #         _explored.append(_sp)
+        # Generate surroundings section File Object Model.
+        _surroundings_settings = SurroundingsSectionFom(
+            surroundings_database_dir=_temp_dir.parent,
+            constructieafstand=float("nan"),
+            constructieovergang=float("nan"),
+            buitendijks=False,
+            bebouwing=False,
+            spoorwegen=False,
+            water=False,
+        )
 
-    @pytest.fixture(name="surroundings_with_obstacle_builder")
-    def _get_surroundings_with_obstacle_builder_fixture(
-        self,
-        distances_to_surrounding_point_builder: Callable[
-            [Point, list[float]], PointSurroundings
-        ],
-    ) -> Iterator[Callable[[list[Point], list[list[float]]], SurroundingsWrapper]]:
-        def create_surroundings_wrapper(
-            point_list: list[Point], distance_list: list[float]
-        ):
-            # _buildings_polderside = SurroundingsObstacle()
-            # _railways_polderside = SurroundingsObstacle()
-            # _waters_polderside = SurroundingsObstacle()
-            # _buildings_polderside.points = list(
-            #     map(
-            #         distances_to_surrounding_point_builder, point_list, distance_list[0]
-            #     )
-            # )
-            # _railways_polderside.points = list(
-            #     map(
-            #         distances_to_surrounding_point_builder, point_list, distance_list[1]
-            #     )
-            # )
-            # _waters_polderside.points = list(
-            #     map(
-            #         distances_to_surrounding_point_builder, point_list, distance_list[2]
-            #     )
-            # )
+        # Generate Infrastructures section file model
+        _infrastructure_settings = InfrastructureSectionFom(
+            infrastructuur=True,
+            opslagfactor_wegen=SurtaxFactorEnum.NORMAAL,
+            infrakosten_0dh=InfraCostsEnum.GEEN,
+            buffer_buitendijks=0.24,
+            wegen_klasse2_breedte=1,
+            wegen_klasse7_breedte=2,
+            wegen_klasse24_breedte=3,
+            wegen_klasse47_breedte=4,
+            wegen_onbekend_breedte=5,
+        )
 
-            # Yield wrapper
-            return SurroundingsWrapper()
+        # Generate wrapper
+        _importer = SurroundingsWrapperCollectionImporter(
+            infrastructure_section_fom=_infrastructure_settings,
+            surroundings_section_fom=_surroundings_settings,
+            selected_locations=["10-1-3-c-1-d-1"],
+            traject_loc_shp_file=_shp_file,
+        )
+        _surroundings_wrapper_list = _importer.build()
+        assert len(_surroundings_wrapper_list) == 1
 
-        yield create_surroundings_wrapper
+        # Yield result
+        yield _surroundings_wrapper_list[0]
 
-    def test_get_obstacle_locations(
-        self, surroundings_with_obstacle_builder: SurroundingsWrapper
-    ):
-        # 1. Define test data.
-        _locations = [
-            Point(4.2, 2.4),
-            Point(4.2, 4.2),
-            Point(2.4, 2.4),
-        ]
-        _distances = [
-            [[], [], []],
-            [[], [], []],
-            [[], [], []],
-        ]
-        _surroundings = surroundings_with_obstacle_builder(_locations, _distances)
-        assert isinstance(_surroundings, SurroundingsWrapper)
-
-        # 2. Run test.
-        _obstacle_locations = _surroundings.obstacle_locations
-
-        # 3. Verify expectations.
-        # Doing it like this because sorting requires setting up a sorting key.
-        assert len(_obstacle_locations) == len(_locations)
-        assert all(isinstance(_ol, PointSurroundings) for _ol in _obstacle_locations)
-
-        _explored = []
-        for _ol_point in _obstacle_locations:
-            assert _ol_point.location in _locations
-            # Check there are no repeated points, thus invalidating the test.
-            assert _ol_point not in _explored
-            _explored.append(_ol_point)
-
-    def test_get_safe_locations(
-        self,
-        surroundings_with_obstacle_builder: Callable[
-            [list[Point], list[list[float]]], SurroundingsWrapper
-        ],
-    ):
-        # 1. Define test data.
-        _locations = [
-            Point(4.2, 2.4),
-            Point(4.2, 4.2),
-            Point(2.4, 2.4),
-        ]
-        _distances = [
-            [[], [20, 50], [10, 50]],
-            [[], [15], [20]],
-            [[5, 100], [5, 100], [5, 100]],
-        ]
-        _surroundings = surroundings_with_obstacle_builder(_locations, _distances)
-        assert isinstance(_surroundings, SurroundingsWrapper)
-
-        # 2. Run test.
-        _safe_points = _surroundings.get_locations_at_safe_distance(12)
-
-        # 3. Verify expectations.
-        # Doing it like this because sorting requires setting up a sorting key.
-        assert all(map(lambda x: isinstance(x, PointSurroundings), _safe_points))
-
-        _explored = []
-        assert len(set(_safe_points)) == len(_locations)
-        for _sp in _safe_points:
-            assert _sp.location in _locations
-            # Check there are no repeated points, thus invalidating the test.
-            assert _sp not in _explored
-            _explored.append(_sp)
+        # Remove temp dir
+        shutil.rmtree(_temp_dir)
 
     @pytest.mark.parametrize(
         "obstacles_distance_list",
@@ -188,6 +113,7 @@ class TestSurroundingsWrapper:
             [list[Point], list[list[float]]], SurroundingsWrapper
         ],
     ):
+        pytest.fail("Needs to be moved to `TestObstacleSurroundingsWrapper`.")
         # 1. Define test data.
         _safe_distance = min(obstacles_distance_list, default=0) - 1
         _wrapper = surroundings_with_obstacle_builder(
@@ -224,69 +150,3 @@ class TestSurroundingsWrapper:
         # 3. Verify expectations.
         assert isinstance(_classified_surroundings, list)
         assert not any(_classified_surroundings)
-
-    @pytest.fixture(name="surroundings_wrapper_with_infrastructure")
-    def _get_surroundings_wrapper_with_infrastructure_fixture(
-        self, request: pytest.FixtureRequest
-    ) -> Iterator[SurroundingsWrapper]:
-        # Shp locations file
-        _shp_file = test_data.joinpath("acceptance", "shp", "dike_locations.shp")
-        assert _shp_file.is_file()
-
-        # Surroundings directory
-        _surroundings_analysis_path = test_data.joinpath(
-            "acceptance", "surroundings_analysis", "10_1"
-        )
-        assert _surroundings_analysis_path.is_dir()
-
-        # Create a dummy dir to avoid importing unnecessary data.
-        _dir_name = get_test_results_dir(request)
-        _temp_dir = test_results.joinpath(_dir_name, "10_1")
-        if _temp_dir.exists():
-            shutil.rmtree(_temp_dir)
-        shutil.copytree(_surroundings_analysis_path, _temp_dir)
-
-        # Generate surroundings section File Object Model.
-        _surroundings_settings = SurroundingsSectionFom(
-            surroundings_database_dir=_temp_dir.parent,
-            constructieafstand=float("nan"),
-            constructieovergang=float("nan"),
-            buitendijks=False,
-            bebouwing=False,
-            spoorwegen=False,
-            water=False,
-        )
-
-        # Generate wrapper
-        _importer = KoswatSurroundingsWrapperCollectionBuilder()
-        _importer.selected_locations = "10-1-3-C-1-D-1"
-        _importer.traject_loc_shp_file = _shp_file
-        _surroundings_wrapper_list = _importer.import_from(_surroundings_settings)
-        assert len(_surroundings_wrapper_list) == 1
-
-        # Yield result
-        yield _surroundings_wrapper_list[0]
-
-        # Remove temp dir
-        shutil.rmtree(_temp_dir)
-
-    def test_get_infrastructure_collection_given_acceptance_case_returns_dict(
-        self, surroundings_wrapper_with_infrastructure: SurroundingsWrapper
-    ):
-        # 1. Define test data
-        assert isinstance(surroundings_wrapper_with_infrastructure, SurroundingsWrapper)
-
-        # 2. Run test.
-        _infra_matrix = (
-            surroundings_wrapper_with_infrastructure.infrastructure_collection
-        )
-
-        # 3. Verify expectations.
-        assert isinstance(_infra_matrix, dict)
-        assert any(_infra_matrix.items())
-        assert all(
-            map(
-                lambda _si: isinstance(_si, SurroundingsInfrastructure),
-                _infra_matrix.values(),
-            )
-        )
